Reorganize package structure and fix imports
