+++ .github/workflows/unit-tests.yml
@@ -24,12 +24,16 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pytest pytest-cov codecov
-        pip install -e math_cli_app/
+        pip install pytest pytest-cov
+        pip install -e "math_cli_app[test]"
+        
+    - name: List directory structure
+      run: |
+        find math_cli_app -type f -name "*.py" | sort
         
     - name: Run tests with pytest
       run: |
-        pytest math_cli_app/unit_test/ --cov=math_cli_app --cov-report=xml -v
+        python -m pytest math_cli_app/unit_test/ -v --cov=math_cli_app --cov-report=xml
         
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
