+++ math_cli_app/math_cli_app/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the math_cli_app directory a Python package
