+++ math_cli_app/unit_test/test_addition.py
@@ -1,5 +1,5 @@
 import pytest
-from addition import add_numbers
+from math_cli_app.addition import add_numbers
 
 def test_add_positive_numbers():
     assert add_numbers(2, 3) == 5
